requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
PyYAML==6.0.1
pytest==8.2.2